+++ core/server/api_container/server/startosis_engine/startosis_interpreter_test.go
@@ -1238,6 +1238,7 @@ func validateScriptOutputFromPrintInstructions(t *testing.T, instructions []kurt
 			require.Nil(t, err, "Error running the print statements")
 			if instructionOutput != nil {
 				scriptOutput.WriteString(*instructionOutput)
+				scriptOutput.WriteString("\n")
 			}
 		}
 	}