+++ internal_testsuites/golang/test_helpers/startosis_helpers.go
@@ -5,11 +5,16 @@ import (
 	"strings"
 )
 
+const (
+	newlineChar = "\n"
+)
+
 func GenerateScriptOutput(instructions []*kurtosis_core_rpc_api_bindings.KurtosisInstruction) string {
 	scriptOutput := strings.Builder{}
 	for _, instruction := range instructions {
 		if instruction.InstructionResult != nil {
 			scriptOutput.WriteString(instruction.GetInstructionResult())
+			scriptOutput.WriteString(newlineChar)
 		}
 	}
 	return scriptOutput.String()